+++ docs/source/prologue/acknowledgements.md
@@ -1,15 +1,21 @@
 # Acknowledgements
 
-## Developer
+## Developers
 
 -   [Isabel
     Pietka](https://bit.ly/2XhuaGQ)  
     isabel.pietka@rub.de
 
-## Contributers
+## Contributors
 
+-   Ralf Drautz
+    ralf.drautz@rub.de
+    
+-   Thomas Hammerschmidt
+    thomas.hammerschmidt@rub.de
 
 
 ## Acknowledgements
 
-
+The authors acknowledge funding by the Deutsche Forschungsgemeinschaft (DFG) through
+project C1 of the collaborative research center SFB/TR 103 (DFG project number 190389738).
