+++ paper/paper.bib
@@ -145,6 +145,6 @@ journal = {Computational Materials Science},
 volume = {68},
 pages = {314-319},
 year = {2013},
-doi = {doi.org/10.1016/j.commatsci.2012.10.028},
+doi = {10.1016/j.commatsci.2012.10.028},
 author = {Shyue Ping Ong and William Davidson Richards and Anubhav Jain and Geoffroy Hautier and Michael Kocher and Shreyas Cholia and Dan Gunter and Vincent L. Chevrier and Kristin A. Persson and Gerbrand Ceder}
 }
