+++ paper/paper.bib
@@ -1,112 +1,97 @@
 @article{Kresse-96,
-title = "Efficiency of ab-initio total energy calculations for metals and semiconductors using a plane-wave basis set",
-journal = "Computational Materials Science",
-volume = "6",
-pages = "15 - 50",
-year = "1996",
-author = "G. Kresse and J. Furthm{\"u}ller",
-doi = {10.1016/0927-0256(96)00008-0}
+	title = "Efficiency of ab-initio total energy calculations for metals and semiconductors using a plane-wave basis set",
+	journal = "Computational Materials Science",
+	volume = "6",
+	pages = "15 - 50",
+	year = "1996",
+	author = "G. Kresse and J. Furthm{\"u}ller",
+	doi = {10.1016/0927-0256(96)00008-0}
 }
 
 @article{Kresse-99,
-title = {From ultrasoft pseudopotentials to the projector augmented-wave method},
-author = {Kresse, G. and Joubert, D.},
-journal = {Physical Review B},
-volume = {59},
-pages = {1758--1775},
-year = {1999},
-doi = {10.1103/PhysRevB.59.1758}
+	title = {From ultrasoft pseudopotentials to the projector augmented-wave method},
+	author = {Kresse, G. and Joubert, D.},
+	journal = {Physical Review B},
+	volume = {59},
+	pages = {1758--1775},
+	year = {1999},
+	doi = {10.1103/PhysRevB.59.1758}
 }
 
 @article{Kresse-96b,
-title = {Efficient iterative schemes for ab initio total-energy calculations using a plane-wave basis set},
-author = {Kresse, G. and Furthm\"uller, J.},
-journal = {Physical Review B},
-volume = {54},
-pages = {11169--11186},
-year = {1996},
-doi = {10.1103/physrevb.54.11169 }
-}
-
-@Article{LAMMPS,
-  author = "A. P. Thompson and H. M. Aktulga and R. Berger and
-     D. S. Bolintineanu and W. M. Brown and P. S. Crozier and
-     P. J. in 't Veld and A. Kohlmeyer and S. G. Moore and T. D. Nguyen and
-     R. Shan and M. J. Stevens and J. Tranchida and C. Trott and S. J. Plimpton",
-  title = "{LAMMPS} - a flexible simulation tool for
-     particle-based materials modeling at the
-     atomic, meso, and continuum scales",
-  journal = "Computer Physics Communications",
-  volume =  "271",
-  pages =   "108171",
-  year =    "2022",
-  doi = "10.1016/j.cpc.2021.108171"
+	title = {Efficient iterative schemes for ab initio total-energy calculations using a plane-wave basis set},
+	author = {Kresse, G. and Furthm\"uller, J.},
+	journal = {Physical Review B},
+	volume = {54},
+	pages = {11169--11186},
+	year = {1996},
+	doi = {10.1103/physrevb.54.11169 }
 }
 
 @article{asr,
-title = {Atomic Simulation Recipes: A Python framework and library for automated workflows},
-journal = {Computational Materials Science},
-volume = {199},
-pages = {110731},
-year = {2021},
-doi = {10.1016/j.commatsci.2021.110731},
-author = {Morten Gjerding and Thorbjørn Skovhus and Asbjørn Rasmussen and Fabian Bertoldo and Ask Hjorth Larsen and Jens Jørgen Mortensen and Kristian Sommer Thygesen}
+	title = {Atomic Simulation Recipes: A Python framework and library for automated workflows},
+	journal = {Computational Materials Science},
+	volume = {199},
+	pages = {110731},
+	year = {2021},
+	doi = {10.1016/j.commatsci.2021.110731},
[…]
