### 2022-06-27 #4519 @thohamm
Missing and invalid DOI fixed.
