### 2022-06-30 #4519 @jedbrown
:wave: Hi @thohamm. A few notes from my initial scan
1. I see there are some "TODO" marks in the document.
2. Can you please share with us a little about the provenance of this project? I see it was a new project importing source from elsewhere just prior to submission.
3. It looks like there is intended to be Sphinx documentation, but I don't see it linked. Does that exist?
4. Those docs say that pytest is used, but it looks like the code doesn't use pytest.
5. Are there unit tests and can they be made to run with continuous integration (such as GitHub Actions)?
6. If the output is tabular, why not put it into a Pandas dataframe for easy plotting and statistics?
7. How does this package related to generic workflow management tools with support for batch computing (e.g., Snakemake, Parsl, Swift, libensemble)?
8. To what extent does this package provide machine- or human-auditable provenance?
9. To what extent does this package convey research, versus serve as a client to research software (VASP, etc.)? Would it be in scope for strucscan to compute summary statistics, make common figures, and/or do quality control to make the research more efficient or reliable?
