### 2022-07-07 5a43e43
Delete projects.md
