+++ docs/source/prologue/projects.md
@@ -1,2 +1,0 @@
-# Projects using strucscan
-
