### 2022-07-07 5c7e2c0
Delete publications.md
