+++ docs/source/prologue/publications.md
@@ -1,3 +1,0 @@
-# Publications using pyscal
-
-
