### 2022-07-07 3636b53
delete pubsandprojects
