### 2022-07-07 1f271cd
Delete methodsindex.rst
