+++ docs/source/methodsindex.rst
@@ -1,6 +1,0 @@
-Methods
--------
-
-.. toctree::
-
-    methods/
