### 2022-07-12 #4519 @thohamm
@danielskatz, @jedbrown - Thank you for your detailed comments.

We reply to them point by point below. We would be willing to include according remarks in the paper if you find this helpful.

1. Thank you for your comment, we fixed them.
2. The development of this python code started about 6 months ago on the basis of 10-year old collection of shell scripts.
   After several iterations of restructuring program flow and data handling we now arrived at a converged version that is also
   already in practical use in our group. We didn't see much benefit in sharing older versions and therefore decided to
   set up a new git repo for this project.
3. The Sphinx documentation has now been prepared and is available on https://strucscan.readthedocs.io/.
4. We prepared tests using pytest but are still struggling with executing them due to write permission problems with GitHub
   Actions. For the moment we removed the corresponding remark from the documentation in the git repo and will include it
   again as soon as the tests are up and running.
6. Some of the examples are planned to be used as unit tests as soon as the issues with github actions are resolved. (cf. 4)
7. The data structure and connectivity of the output depends on the particular set of calculations and is therefore not very well
   suited for a pandas dataframe with an a-priori layout. We are instead storing dictionaries from which the user can collect
   data selectively for constructing dedicated dataframes. This is more convenient and flexible in our experience.
8. In contrast to generic tools for batch computing, strucscan is dedicated to high-throughput simulations and provides a more
   sophisticated framework than a generic workflow manager. In particular, it offers generalized interfaces to different
   simulation software packages, interfaces to schedulers including monitoring tools, a transparent storage structure for
   convenient post-processing, and measures for data connectivity.
9. The provenance of the strucscan code itself is given by the github versioning that starts at the first converged version of the
   software (cf. 2). The provenance of individual calculations with strucscan is realized by storing all input and output files in
   a folder tree. This storage is in addition to the collection of central results in dictionaries. This approach of storing the full
   data set as provenance of individual calculations provides the user full flexibility of adapting to external meta-data schema,
   e.g. of  third-party databases.
10. The goal of strucscan is to support the user to convey high-throughput research. This means particularly to cope with the
   challenges of starting, monitoring, collecting and analyzing very large numbers (thousands, ten-thousands) of calculations.
   The goal of strucscan is to provide the basis for exactly the mentioned purposes, i.e.,  summary statistics, common figures
   and quality control. This means a framework for generating large sets of quality-checked data stored in a transparent data
   structure. On this ground the user can then use the basic post-processing of strucscan or further tools of data visualization
   and analytics.
