+++ setup.py
@@ -25,18 +25,18 @@ setup(
 
     classifiers=['Development Status :: 5 - Production/Stable',
                  'Topic :: Scientific/Engineering :: Physics',
-                 ''License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
+                 'License :: OSI Approved :: GNU General Public License v3 (GPLv3)',
                  'Intended Audience :: Science/Research',
                  'Operating System :: OS Independent',
                  'Programming Language :: Python :: 3',
-                 'Programming Language :: Python :: 3.4',
-                 'Programming Language :: Python :: 3.5',
                  'Programming Language :: Python :: 3.6',
-                 'Programming Language :: Python :: 3.7'],
+                 'Programming Language :: Python :: 3.7',
+                 'Programming Language :: Python :: 3.8',
+                 'Programming Language :: Python :: 3.9'],
 
     keywords='computational materials science',
     packages=find_packages(include=['strucscan', 'strucscan.*']),
-    install_requires=['ase', 'numpy', 'pyyaml'],
+    install_requires=['ase', 'numpy'],
     cmdclass=versioneer.get_cmdclass(),
     include_package_data=True,
     package_data={'': extra_files},
