+++ docs/requirements.txt
@@ -1,5 +1,6 @@
 numpy
 ase
+pyyaml
 ipywidgets==7.5
 sphinx_rtd_theme
 m2r2
