### 2022-08-16 #4519 @thohamm
@jedbrown Thank you for your comments!

- We included pyyaml in the installation setup for pip and conda-forge.
- We added a brief print out page when calling strucscan with '--help'.
- Regarding the scope we would like to emphasize that we developed strucscan as tool for conducting, monitoring and collecting data from high-throughput atomistic simulation calculations. We included an interface to a widely used DFT code (VASP) as example and provide a simple interface for extension to other codes. We know from our  own experience and from discussions with other groups that there is a need for transparent and easily-extendable high-throughput tools for atomistic simulations. We are therefore convinced that strucscan will be useful beyond our group and will support users to adapt it for  their needs.
