### 2022-08-31 #4719 @editorialbot
[REVIEW]: strucscan: A lightweight Python-based framework for high-throughput material simulation
**Submitting author:** @thohamm (<a href="http://orcid.org/0000-0002-2270-4469">Thomas Hammerschmidt</a>)
**Repository:** https://github.com/ICAMS/strucscan
**Branch with paper.md** (empty if default branch):
**Version:** 1.0
