### 2022-09-11 ICAMS/strucscan#41 @mturiansky
JOSS Review - unit tests
As part of the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4719), unit tests are required to demonstrate the functionality of the code:

> Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?

While _automated_ tests are not required, I would highly suggest implementing them using github actions as it will help with the long-term maintainability of the code. Similarly, I would recommend setting up a linter (but that's optional of course).

I see that there is something implemented in `tests/`, but it includes hard-coded paths, which are specific to the maintainer's system. These fail when I run it on my computer.
