### 2022-09-11 ICAMS/strucscan#42 @mturiansky
JOSS Review - docs
[JOSS Review](https://github.com/openjournals/joss-reviews/issues/4719)

I see that almost all of the implemented methods include docstrings, but the [API documentation](https://strucscan.readthedocs.io/en/latest/modules.html) does not seem to contain the same information and is a bit lacking. It should be straightforward to update the API documentation with the implemented docstrings and [sphinx-apidoc](https://www.sphinx-doc.org/en/master/man/sphinx-apidoc.html).
