### 2022-09-11 ICAMS/strucscan#43 @mturiansky
JOSS Review - functionality
I am trying to test the functionality of your code as part of the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/4719), but I am running into trouble with running it. I get the following error:

```
» strucscan examples/dummy.yaml
Data tree path:                /home/mark/Downloads/strucscan/data
Structure repository:          /home/mark/Downloads/strucscan/structures
Resource repository:           /home/mark/Downloads/strucscan/resources

Optional key 'monitor' not provided. Default value will be used: True
Optional key 'submit' not provided. Default value will be used: True
Optional key 'collect' not provided. Default value will be used: True
Traceback (most recent call last):
  File "/home/mark/.virtualenvs/tmp-54286c793853f99/bin/strucscan", line 33, in <module>
    sys.exit(load_entry_point('strucscan', 'console_scripts', 'strucscan')())
  File "/home/mark/Downloads/strucscan/strucscan/cli.py", line 31, in main
    JobManager(input_dict)
  File "/home/mark/Downloads/strucscan/strucscan/core/jobmanager.py", line 118, in __init__
    if len(properties) > 1:
NameError: name 'properties' is not defined. Did you mean: 'property'?
```

I've tried a few different parameters in `~/.strucscan` and the input file but to no avail. Any help would be appreciated.
