### 2022-09-23 0169199
set default PROJECT_PATH
