### 2022-10-10 4a002a5
Delete config.yaml
