### 2022-10-10 37c429c
Delete resources/machineconfig/zgh directory
