### 2022-10-10 deae128
Delete testing.yml
