+++ .github/workflows/testing.yml
@@ -1,42 +1,0 @@
-name: strucscan-testing
-
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
-
-jobs:
-  build:
-    strategy:
-      matrix:
-        os: [ubuntu-latest, macos-latest]
-        python-version: [3.6, 3.7, 3.9]
-
-    runs-on: ${{ matrix.os }}
-    steps:
-    - uses: actions/checkout@v2
-    - name: Cache conda
-      uses: actions/cache@v1
-      env:
-        # Increase this value to reset cache if etc/example-environment.yml has not changed
-        CACHE_NUMBER: 0
-      with:
-        path: ~/conda_pkgs_dir
-        key:
-          ${{ runner.os }}-conda-${{ env.CACHE_NUMBER }}-${{
-          hashFiles('.ci_support/environment.yml') }}
-    - uses: conda-incubator/setup-miniconda@v2
-      with:
-        activate-environment: py3
-        channel-priority: strict
-        environment-file: .ci_support/environment.yml
-        python-version: ${{ matrix.python-version }}
-        use-only-tar-bz2: true # IMPORTANT: This needs to be set for caching to work properly!
-
-    - name: run tests
-      shell: bash -l {0}
-      run: |
-        pip install -e .
-        pip install pytest
-        pytest tests/
