+++ .github/workflows/github-actions-deml.yml
@@ -1,10 +1,12 @@
 name: strucscan-testing
+run-name: ${{ github.actor }} is testing
 
-on:
-  push:
-    branches: [ master ]
-  pull_request:
-    branches: [ master ]
+on: [push]
+#on:
+#  push:
+#    branches: [ master ]
+#  pull_request:
+#    branches: [ master ]
 
 jobs:
   build:
