### 2022-10-10 ICAMS/strucscan#62 @mturiansky
JOSS Review - functionality
[Joss Review](https://github.com/openjournals/joss-reviews/issues/4719)

The code now runs for me on my computer and finishes, but I see a series of errors like the following:
```
>> collecting DUMMY/Al/total__fcc__Al ...
Problem in collecting 'total' properties from /.../strucscan//DUMMY/Al/total__fcc__Al:
line 31 in collector.py : <class 'RuntimeError'>
['  File "/.../strucscan/strucscan/core/collector.py", line '
 '31, in get_result_dict\n'
 '    result_dict = bulk.get_bulk_properties(calc, absolute_path)\n',
 '  File "/.../strucscan/strucscan/properties/bulk.py", line '
 '8, in get_bulk_properties\n'
 '    energy = final_struct.get_potential_energy()\n',
 '  File '
 '"/.../tmp-cbb0fa3f4c0dc28/lib/python3.10/site-packages/ase/atoms.py", '
 'line 726, in get_potential_energy\n'
 "    raise RuntimeError('Atoms object has no calculator.')\n"]
Atoms object has no calculator.
```

Is this an expected side-effect of running the dummy job? Or is this some error in strucscan itself? I find it hard to discern based on the error that is being output.
