### 2022-10-10 #4719 @mturiansky
I opened one last issue. Once this is addressed, I can tick off the functionality checkbox and am okay with publication. The test coverage is a bit low in my opinion (see below), but I think it does cover the basic functionality.

```---------- coverage: platform linux, python 3.10.7-final-0 -----------
Name                                Stmts   Miss  Cover
-------------------------------------------------------
strucscan/__init__.py                  10      0   100%
strucscan/_version.py                 337    213    37%
strucscan/cli.py                       35     35     0%
strucscan/core/__init__.py              0      0   100%
strucscan/core/collector.py            28      0   100%
strucscan/core/datatree.py             37      4    89%
strucscan/core/jobmaker.py            138     66    52%
strucscan/core/jobmanager.py          252     61    76%
strucscan/core/jobobject.py            50      5    90%
strucscan/core/statusmanager.py        28     11    61%
strucscan/engine/__init__.py            0      0   100%
strucscan/engine/dummy.py             115     70    39%
strucscan/engine/generalengine.py      63     22    65%
strucscan/engine/vasp.py              244    215    12%
strucscan/error/__init__.py             0      0   100%
strucscan/error/errorhandler.py       107     93    13%
strucscan/error/errormanager.py        27     24    11%
strucscan/properties/__init__.py        0      0   100%
strucscan/properties/bulk.py           11      8    27%
strucscan/properties/eos.py            41     25    39%
strucscan/resources/__init__.py         0      0   100%
strucscan/resources/atomicdata.py       2      0   100%
strucscan/resources/inputyaml.py       22      0   100%
strucscan/resources/properties.py      32      5    84%
strucscan/scheduler.py                168    119    29%
strucscan/utils.py                    236    108    54%
-------------------------------------------------------
TOTAL                                1983   1084    45%
```
