### 2022-10-11 826597b
add empty SinglePointCalculator to final_struct
