### 2022-10-11 65b3280
import numpy
