### 2022-10-11 ICAMS/strucscan#62 @thohamm
Thank you for reporting the error. Indeed, it comes from the fact that the dummy engine is implemented for test purposes only and tests file creation, writing and reading. We added the assignment of an ASE SinglepointCalculator object to the final structures ('final_struct') that assigns the structure a energy of value zero. This fixes the output of the test.
