### 2022-10-20 ICAMS/strucscan#63 @wcwitt
JOSS review - Quality of writing
As part of [the review](https://github.com/openjournals/joss-reviews/issues/4719), I'm required to verify that the paper is well-written, and it is. However, the quality of writing in the documentation is a bit lower, with a fairly high number of typos and confusing sentences.

I started to list some typos, but decided instead to just fix them in a PR. I know I didn't find everything, so please review the documentation and fix anything I've missed.
