### 2022-10-20 #4719 @wcwitt
I've now managed to install and run the Dummy examples, both from the Jupyter notebook and the command line. I don't have a VASP license, so I can't verify those parts. Overall, I think the submission is of good quality. There are two issues (linked above) for the authors to respond to.
