### 2022-10-20 4e322aa
add info on config file
