### 2022-10-20 96d94d2
add info on resource directory
