### 2022-10-20 2aef54a
add 05_add_new_properties
