### 2022-10-20 03983a4
add resource repo example
