+++ examples/structures/unaries/bulk/fcc.cfg
@@ -1,0 +1,16 @@
+Number of particles = 1
+A = 1.0 Angstrom
+H0(1,1) = 0.000000 A
+H0(1,2) = 2.010000 A
+H0(1,3) = 2.010000 A
+H0(2,1) = 2.010000 A
+H0(2,2) = 0.000000 A
+H0(2,3) = 2.010000 A
+H0(3,1) = 2.010000 A
+H0(3,2) = 2.010000 A
+H0(3,3) = 0.000000 A
+.NO_VELOCITY.
+entry_count = 3
+1.008000
+H
+0.000000e+00 0.000000e+00 -0.000000e+00 
