### 2022-10-20 0b49ead
add info on structure repo
