+++ docs/source/prologue/extending.md
@@ -12,22 +12,21 @@ In case a bug is found in the module, it can be reported on the [issues
 page of the 
 repository](https://github.com/ICAMS/strucscan/issues). 
 After clicking on the new issue button, there is template already provided for
-Bug report. Please choose this and make sure the necessary fields are
+bug report. Please choose this and make sure the necessary fields are
 filled. Once a bug is reported, the status can once again monitored on
 the issues page. Additionally, you are of course very welcome to fix any
 existing bugs.
 
 ## New features
 
-If you have an idea for new feature, you can submit a feature idea
+If you have an idea for a new feature, you can submit a feature idea
 through the [issues page of the
 repository](https://github.com/ICAMS/strucscan/issues). After choosing
-new issue, please choose the template for feature request. As much as
-information as you can provide about the new feauture would be greatly
-helpful. Additionally, you could also work on feature requests already
-on the issues page. 
+new issue, please choose the template for feature request. Detailed
+information about the new feauture will be very helpful. Additionally, 
+you can work on feature requests on the issues page. 
 If you have trouble, or you need help, please
 [send an email](mailto:isabel.pietka@rub.de) and we will be happy to
 help! All of the contributions are greatly appreciated and will be
-credited in Developers/Acknowledgements page.
+credited in the Developers/Acknowledgements section.
 
