### 2022-10-21 ICAMS/strucscan#63 @thohamm
Thank you for opening this issue. We fixed typos and clarified sentences in the documentation and examples. We updated the git repo and closed the related PR.
