### 2022-10-24 #4719 @ppxasjsm
Thank you! @thohamm can you go through the PDF of the paper version and do a check to make sure you are happy with the current state? I'll also go through it one last time, but won't likely get a chance to do so until next week I am afraid.
