### 2022-11-11 #4719 @thohamm
Thank you very much, we are glad that the paper is accepted. The latest version in the git repo is now released with tag v1.0.0 and uploaded to zenodo. DOI: 10.5281/zenodo.7313602
