### 2022-11-23 #4719 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
