### 2022-11-27 #4719 @Kevin-Mattheus-Moerman
@ppxasjsm thanks for editing this submission. A minor feedback point. Please include the `v` if it is part of the version tag e.g. `v1.0.0`. So the version tag should match the one shown here: `https://github.com/ICAMS/strucscan/tags`. I'll also ask the authors to edit the version listed on ZENODO which currently says `version 1.0.0`.
