### 2022-11-27 #4719 @Kevin-Mattheus-Moerman
@thohamm I am an AEiC for this track and here to help process acceptance of this work in JOSS. I've studied your repository, the archived version, and this review. I've also proofread your paper. I have the following minor points that require your attention before we proceed to process this for acceptance (feel free to disagree with language suggestions or to propose your own alternatives if needed):

- [ ] I have set the version tag as listed on your repository i.e. `v1.0.0`, can you please manually edit the ZENODO archive to list this as the version, it currently says `version 1.0.0`.
- [ ] The paper contains both `data tree` and `data-tree`, please uses the latter throughout.
- [ ] Use past tense in `...calculations in computational materials science lead to the...` i.e. `...calculations in computational materials science have lead to the...`
- [ ] Commas are needed in the locations indicated by square brackets: `...can be extended[,] with basic programming skills[,] to further scheduling systems...`. And I suggest actually rephrasing this sentence to the smoother (check if I understood it right, the lack of commas confused me): `...can be extended, by somebody with basic programming skills, with additional scheduling systems, simulation codes and material properties at the atomic scale, and well as other simulation scales.`. Feel free to propose a different alternative.
- [ ] Commas needed at square bracket locations: `...a large toolbox of features[,] or from a predefined database concept[,] but rather needs...`
- [ ] Should `materials properties` be written `material properties`?

Recommendations (not required):
- [ ] If the acronym in the affiliation `(ICAMS)` is not used elsewhere in the document, you could consider removing it.
- [ ] Add commas as shown by square bracket locations
`...by computational material science relies[,] to a large degree[,] on the prediction...`
`...different crystal structures[,] or different simulation settings.`
- [ ] Consider using `:` after `are` in `...the required high-throughput calculations are (i) the sampling ...`
- [ ] It is common practice to use code style text, when referring to software project names. So instead of saying strucscan you may consider using `strucscan`. i.e. use:
`strucscan`
```
If you do this, also do it for `SunGridEngine` and `slurm` and any other software project names.
- [ ] Remove the capital letter when strucscan appears at the start of a sentence, as the project name does not feature one.
- [ ] The text mentions `Dependencies:` as a "feature", however you mean the opposite, i.e. the fact that your package has few dependencies is a bonus. So perhaps consider a different word, or something like `Reduced dependencies`.
