### 2022-11-28 #4719 @thohamm
@Kevin-Mattheus-Moerman Thank you for your careful reading and suggestions.

The following points are changed:
- set zenodo version to v1.0.0
- changed to 'data-tree' throughput
- inserted past tense
- inserted all three comma suggestions and simplified sentence 'can be extended ..'
- changed to 'strucscan' in lowercase throughput

The remaining points were not changed as they are common terms or fine in our opinion.
