### 2022-11-28 #4719 @Kevin-Mattheus-Moerman
@thohamm congratulations for this publication in JOSS!

@ppxasjsm thanks for editing!

Special thanks to the reviewers @mturiansky and @wcwitt!! :tada:
