### 2022-11-28 #4719 @ppxasjsm
> @editorialbot set v1.0.0 as version

@Kevin-Mattheus-Moerman Thank you! This was my first one. I'll remember that next time.
