### 2022-11-28 #4719 @thohamm
Thank you all for your efforts in editing, reviewing and publishing our paper. We are glad that our work could be published in JOSS!
